requirements.txt
numpy==1.19.5
scipy==1.5.4
PyYAML==5.4.1
tensorly==0.5.1
tqdm==4.59.0
pandas==1.1.5
torch==1.6.0
torchvision==0.7.0
